Teamwork-5/requirements.txt
asgiref==3.7.1
dj-rest-auth==4.0.1
Django==4.2.1
djangorestframework==3.14.0
Pillow==9.5.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
